+++ Version/VersionInfo.cs
@@ -5,7 +5,7 @@ using System.Resources;
 [assembly: AssemblyProduct("RVis")]
 [assembly: AssemblyCopyright("Copyright © HSE 2020")]
 
-[assembly: AssemblyVersion("0.9.11068.1")]
-[assembly: AssemblyFileVersion("0.9.11068.1")]
+[assembly: AssemblyVersion("0.9.11068.2")]
+[assembly: AssemblyFileVersion("0.9.11068.2")]
 
 [assembly: NeutralResourcesLanguage("en")]